add hero title, subtitle and themeselector
alpha v6
