+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@markdownspace/components",
   "private": false,
-  "version": "0.0.1-alpha.5",
+  "version": "0.0.1-alpha.6",
   "type": "module",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
