+++ src/index.ts
@@ -8,7 +8,7 @@ export { Dropdown } from "./components/Dropdown";
 export { File } from "./components/File";
 export { Form } from "./components/Form";
 export { FormField } from "./components/FormField";
-export { Hero } from "./components/Hero";
+export { Hero, HeroTitle, HeroSubtitle } from "./components/Hero";
 export { HeroSection } from "./components/HeroSection";
 export { Icon, IconText } from "./components/Icon";
 export { Image } from "./components/Image";
@@ -33,3 +33,4 @@ export { Tag } from "./components/Tag";
 export { Textarea } from "./components/Textarea";
 export { Title } from "./components/Title";
 export { Typography } from "./components/Typography";
+export { ThemeSelector } from "./components/ThemeSelector";
